Rename partial -> inline for components.
