+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/shop/layout/inline/AmountRenderComponent.java
@@ -1,4 +1,4 @@
-package com.ghostchu.quickshop.shop.layout.partial;
+package com.ghostchu.quickshop.shop.layout.inline;
 
 /*
  * QuickShop-Hikari
@@ -22,23 +22,22 @@ import com.ghostchu.quickshop.api.localization.text.ProxiedLocale;
 import com.ghostchu.quickshop.api.shop.Shop;
 import com.ghostchu.quickshop.api.shop.SignRenderSnapshot;
 import com.ghostchu.quickshop.api.shop.layout.RenderComponent;
-import com.ghostchu.quickshop.util.Util;
 import net.kyori.adventure.text.Component;
 import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 
 /**
- * ItemNameRenderComponent
+ * AmountRenderComponent
  *
  * @author creatorfromhell
  * @since 6.3.0.0
  */
-public class ItemNameRenderComponent implements RenderComponent {
+public class AmountRenderComponent implements RenderComponent {
 
   @Override
   public String placeholder() {
 
-    return "<item_name>";
+    return "<amount>";
   }
 
   @Override
@@ -56,12 +55,12 @@ public class ItemNameRenderComponent implements RenderComponent {
   @Override
   public Component renderSnapshot(final @NotNull SignRenderSnapshot snapshot, final ProxiedLocale locale) {
 
-    return snapshot.itemName();
+    return Component.text(snapshot.itemAmount());
   }
 
   @Override
-  public Component render(final @NotNull Shop shop, @NotNull final ItemStack item, final ProxiedLocale locale) {
+  public Component render(final @NotNull Shop shop, final @NotNull ItemStack item, final ProxiedLocale locale) {
 
-    return Util.getItemStackName(item, locale.getLocale());
+    return Component.text(item.getAmount());
   }
 }
