+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/shop/layout/inline/LevelRenderComponent.java
@@ -1,4 +1,4 @@
-package com.ghostchu.quickshop.shop.layout.partial;
+package com.ghostchu.quickshop.shop.layout.inline;
 
 /*
  * QuickShop-Hikari
@@ -18,26 +18,33 @@ package com.ghostchu.quickshop.shop.layout.partial;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
+import com.ghostchu.quickshop.QuickShop;
 import com.ghostchu.quickshop.api.localization.text.ProxiedLocale;
 import com.ghostchu.quickshop.api.shop.Shop;
 import com.ghostchu.quickshop.api.shop.SignRenderSnapshot;
 import com.ghostchu.quickshop.api.shop.layout.RenderComponent;
+import com.ghostchu.quickshop.util.Util;
 import net.kyori.adventure.text.Component;
+import org.bukkit.enchantments.Enchantment;
 import org.bukkit.inventory.ItemStack;
+import org.bukkit.inventory.meta.FireworkMeta;
+import org.bukkit.potion.PotionEffect;
 import org.jetbrains.annotations.NotNull;
 
+import java.util.Map;
+
 /**
- * AmountRenderComponent
+ * LevelRenderComponent
  *
  * @author creatorfromhell
  * @since 6.3.0.0
  */
-public class AmountRenderComponent implements RenderComponent {
+public class LevelRenderComponent implements RenderComponent {
 
   @Override
   public String placeholder() {
 
-    return "<amount>";
+    return "<item_level>";
   }
 
   @Override
@@ -55,12 +62,32 @@ public class AmountRenderComponent implements RenderComponent {
   @Override
   public Component renderSnapshot(final @NotNull SignRenderSnapshot snapshot, final ProxiedLocale locale) {
 
-    return Component.text(snapshot.itemAmount());
+    return snapshot.levelLine();
   }
 
   @Override
-  public Component render(final @NotNull Shop shop, final @NotNull ItemStack item, final ProxiedLocale locale) {
+  public Component render(final @NotNull Shop shop, @NotNull final ItemStack item, final ProxiedLocale locale) {
 
-    return Component.text(item.getAmount());
+    final QuickShop plugin = QuickShop.getInstance();
+
+    final PotionEffect effect = Util.getFirstPotionEffect(item);
+    if(effect != null) {
+
+      return plugin.text().of("signs.item-duration", Util.getPotionDuration(effect)).forLocale(locale.getLocale());
+    }
+
+    final Map.Entry<Enchantment, Integer> enchantment = Util.getFirstEnchantment(item);
+    if(enchantment != null) {
+
+      return plugin.text().of("signs.item-level", enchantment.getValue()).forLocale(locale.getLocale());
+    }
+
+    //TODO: Parser API? Tying rendering to specific parts, or adding custom rendering logic
+    if(item.getItemMeta() instanceof final FireworkMeta fireworkMeta) {
+
+      return plugin.text().of("signs.item-duration", fireworkMeta.getPower()).forLocale(locale.getLocale());
+    }
+
+    return Component.empty();
   }
 }
