+++ quickshop-bukkit/src/main/java/com/ghostchu/quickshop/shop/layout/inline/OwnerRenderComponent.java
@@ -1,4 +1,4 @@
-package com.ghostchu.quickshop.shop.layout.partial;
+package com.ghostchu.quickshop.shop.layout.inline;
 
 /*
  * QuickShop-Hikari
@@ -18,7 +18,6 @@ package com.ghostchu.quickshop.shop.layout.partial;
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import com.ghostchu.quickshop.QuickShop;
 import com.ghostchu.quickshop.api.localization.text.ProxiedLocale;
 import com.ghostchu.quickshop.api.shop.Shop;
 import com.ghostchu.quickshop.api.shop.SignRenderSnapshot;
@@ -28,17 +27,17 @@ import org.bukkit.inventory.ItemStack;
 import org.jetbrains.annotations.NotNull;
 
 /**
- * PriceAloneRenderComponent
+ * OwnerRenderComponent
  *
  * @author creatorfromhell
  * @since 6.3.0.0
  */
-public class PriceAloneRenderComponent implements RenderComponent {
+public class OwnerRenderComponent implements RenderComponent {
 
   @Override
   public String placeholder() {
 
-    return "<price_alone>";
+    return "<owner>";
   }
 
   @Override
@@ -56,12 +55,12 @@ public class PriceAloneRenderComponent implements RenderComponent {
   @Override
   public Component renderSnapshot(final @NotNull SignRenderSnapshot snapshot, final ProxiedLocale locale) {
 
-    return Component.text(snapshot.formattedPrice());
+    return snapshot.ownerName();
   }
 
   @Override
   public Component render(final @NotNull Shop shop, final @NotNull ItemStack item, final ProxiedLocale locale) {
 
-    return Component.text(QuickShop.getInstance().getShopManager().format(shop.getPrice(), shop));
+    return shop.ownerName();
   }
 }
